requirements.txt
flake8==7.3.0
playwright==1.58.0
pytest==9.0.3
python-dotenv==1.2.2

faker
allure-python-commons